Implement more generic MultiVersion plugin.
Versions are autodetected now. As long as a Resource of the name MVP/{version} exists in the jar it will load classes under {version} as if they are in the root based upon the server version. It will also look for newer versions in natural order to determine how to load classes.